--- jest.config.js
@@ -1,5 +1,0 @@
-// eslint-disable-next-line strict
-module.exports = {
-    preset: 'ts-jest',
-    testEnvironment: 'node',
-};
