+++ package.json
@@ -9,8 +9,6 @@
     "dev": "nodemon --watch src/ -e ts --exec ts-node src/index.ts",
     "run": "npm run build:typescript && npm run run:node",
     "watch": "-tsc -w",
-    "test": "jest",
-    "test:watch": "jest --watch",
     "lint": "eslint src/**/**/*.ts",
     "lint-fix": "eslint src/**/**/*.ts --fix"
   },
@@ -31,15 +29,12 @@
   },
   "devDependencies": {
     "@heavy-division/eslint-config": "^0.1.0",
-    "@types/jest": "^28.1.6",
     "@types/node": "^18.0.0",
     "@typescript-eslint/eslint-plugin": "^4.28.5",
     "@typescript-eslint/parser": "^4.28.5",
     "dotenv": "^16.0.0",
     "eslint": "^7.29.0",
-    "jest": "^28.1.3",
     "nodemon": "^2.0.19",
-    "ts-jest": "^28.0.7",
     "ts-node": "^10.4.0",
     "typescript": "^4.3.5"
   }
